Move the selection API to another file
